move util, fix test
